requirements.txt
django~=4.1.3
djangorestframework~=3.14.0
djoser

requests~=2.28.1
PyYAML~=6.0